AI-Model/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
python-dotenv
pydantic
python-multipart
pypdf2